+++ ViewModels/AccountAddAccountViewModel.cs
@@ -6,6 +6,8 @@ namespace IllinoisProject.ViewModels
     public class AccountAddAccountViewModel
     {
         public Account Account { get; set; }
+        public SelectList AccountList { get; set; }
+        public BlogPost BlogPost { get; set; }
      //   public Major Major { get; set; }
         //public SelectList? MajorList { get; set; }
         //commented out line 10 because I wasn't sure if it was suppose to be there
