+++ Views/BlogPost/AllBlogPost.cshtml
@@ -15,6 +15,10 @@
             <tr>
             <th>Blog name</th>
             <th>Blog Description</th>
+             <th>Account Name</th>
+             <th>Edit</th>
+             <th>Delete</th>
+
             
             </tr>
             @{
@@ -23,9 +27,10 @@
                     <tr>
                         <td>@blogPost.BlogName</td>
                         <td>@blogPost.BlogDescription</td>
+                        <td>@blogPost.Account.AccountName</td>
                        
 
-                        <td><a asp-action="EditBlog" asp-route-id="@blogPost.BlogDescription">Edit</a></td>
+                        <td><a asp-action="EditBlog" asp-route-id="@blogPost.BlogPostId">Edit</a></td>
                         
                         <td><a asp-action="DeleteBlog" asp-route-id="@blogPost.">Delete</a></td>
